+++ gruntfile.js
@@ -1,37 +1,74 @@
 module.exports = function(grunt) {
-	grunt.loadNpmTasks('grunt-nuget');
-	grunt.loadNpmTasks('grunt-contrib-requirejs');
-  	grunt.loadNpmTasks('grunt-contrib-cssmin');
-	grunt.loadNpmTasks('grunt-jsdoc');
-	grunt.loadNpmTasks('grunt-contrib-clean');
-
+	require('load-grunt-tasks')(grunt);
+	var dist = 'dist';
+	var VERSION_REGEXP = /\bv(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)\.(?:0|[1-9][0-9]*)(?:-[\da-z\-]+(?:\.[\da-z\-]+)*)?(?:\+[\da-z\-]+(?:\.[\da-z\-]+)*)?\b/i;
 	grunt.initConfig({
 		pkg: grunt.file.readJSON('package.json'),
-		requirejs: {
-			compile: {
-				options: {
-					baseUrl: "src/js",
-					name: "ajax-alerts",
-					out: "dist/js/ajax-alerts.min.js",
-					optimize: "uglify2",
-					preserveLicenseComments: true,
-					paths: {
-						'jquery': 'empty:'
-					}
-				}
+		banner: '/*! <%= pkg.description %>\n *\n' +
+		' * <%= pkg.name %> v<%= pkg.version %> (<%= pkg.homepage %>)\n' +
+		' * Copyright 2013-<%= grunt.template.today("yyyy") %> <%= pkg.author %>\n' +
+		' * Licensed under <%= pkg.license.type %> (<%= pkg.license.url %>)\n *\n' +
+		' */\n',
+		usebanner: {
+		    dist: {
+		      	options: {
+			        position: 'top',
+			        banner: '<%= banner %>'
+			    },
+			    files: {
+			        src: [ dist + '/js/*.js', dist + '/css/*.css' ]
+			    }
+		    }
+		},
+		bump: {
+			options: {
+				files: [ 'bower.json', 'package.json', 'src/js/<%=pkg.name%>.js' ],
+				updateConfigs: [ 'pkg' ],
+				commit: false,
+				createTag: false,
+				tagName: '%VERSION%',
+				tagMessage: '%VERSION%',
+				push: false,
+				globalReplace: true
 			}
 		},
-
+		uglify: {
+			dist: {
+				options: {
+					// banner: '<%= banner %>'
+				},
+				src: 'src/js/<%= pkg.name %>.js',
+				dest: dist + '/js/<%= pkg.name %>.min.js'
+			}
+		},
+		replace: {
+			dist: {
+			    src: [ dist + '/js/*.js'],
+			    dest: dist + '/js/',
+			    replacements: [{
+			      from: '?define',
+			      to: '? define' // dojo loader has issue without space. requirejs works fine
+			    }]
+			},
+			src: {
+				src: [ 'src/js/<%=pkg.name%>.js'],
+				dest: [ 'src/js/<%=pkg.name%>.js'],
+				replacements: [{
+					from: VERSION_REGEXP,
+					to: 'v<%=pkg.version%>'
+				}]
+			}
+		},
 		cssmin: {
 	      single: {
 	        files: {
-	          'dist/css/ajax-alerts.min.css': ['src/css/ajax-alerts.css']
+	          'dist/css/<%= pkg.name %>.min.css': ['src/css/<%= pkg.name%>.css']
 	        }
 	      }
 	    }, 
 	    jsdoc: {
 	      dist: {
-	        src: ['./src/js/ajax-alerts.js', './readme.md'],
+	        src: ['./src/js/<%= pkg.name%>.js', './readme.md'],
 	        options: {
 	          destination: './docs',
 	          configure: './jsdoc.conf.json',
@@ -54,6 +91,6 @@ module.exports = function(grunt) {
 	});
 	grunt.registerTask('pack', ['nugetpack']);
   	grunt.registerTask('doc', ['clean:docs', 'jsdoc']);
-  	grunt.registerTask('build', ['requirejs', 'cssmin']);
-	grunt.registerTask('default', ['requirejs', 'cssmin', 'clean:docs', 'jsdoc', 'nugetpack']);
+  	grunt.registerTask('build', ['uglify', 'replace', 'cssmin', 'usebanner']);
+	grunt.registerTask('default', ['uglify', 'replace', 'cssmin', 'usebanner', 'clean:docs', 'jsdoc', 'nugetpack']);
 }